urlshare/requirements.txt
django
django-imagekit 
Pillow